requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
astroid==2.15.8
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
billiard==3.6.4.0
black==23.1.0
cachetools==5.3.3
celery==5.2.7
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.2.1
crispy-bootstrap5==0.7
cron-descriptor==1.4.3
dataclasses-json==0.6.6
decorator==5.1.1
dill==0.3.8
distlib==0.3.8
Django==4.2.13
django-celery-beat==2.5.0
django-coverage-plugin==3.0.0
django-crispy-forms==2.0
django-debug-toolbar==3.8.1
django-environ==0.10.0
django-extensions==3.2.1
django-model-utils==4.3.1
django-redis==5.2.0
django-stubs==1.15.0
django-stubs-ext==5.0.0
django-timezone-field==6.1.0
docutils==0.19
exceptiongroup==1.2.1
executing==2.0.1
factory-boy==3.2.1
Faker==25.1.0
filelock==3.14.0
flake8==6.0.0
flake8-isort==6.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.0
hyperframe==6.0.1
identify==2.5.36
idna==3.7
imagesize==1.4.1
importlib-metadata==7.1.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.18.1
isort==5.13.2
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
loguru==0.7.2
numpy==1.26.4
orjson==3.10.3
packaging==23.2
pandas==2.0.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
Pillow
platformdirs==4.2.1
pluggy==1.5.0
portalocker==2.8.2
pre-commit==3.1.1
prompt-toolkit==3.0.43
psycopg2-binary
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.7.1
pydantic-core==2.18.2
pyflakes==3.0.1
pygments==2.18.0
pylint==2.17.7
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyparsing==3.1.2
pytest==7.2.2
pytest-django==4.5.2
pytest-sugar==0.9.6
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.1
whitenoise==6.6.0
gunicorn
